Remaining application code also moving repository
The remaining application-level code is moving to the samtools project,
in the samtools/samtools repository.  A duplicate of bamshuf.c is already
there; related existing functionality there will be extended to provide
the functionality provided by bam2bed.c, bam2fq.c, and bamidx.c; and
abreak.c will be included in samtools as appropriate.