Replace HTS_VERSION with hts_version()
Provide a function returning the htslib version number, as the
interesting version number for user code is that of the library
it is running against, rather than the one it was built against.

Burning the version number into the library object (rather than some
arrangement of headers and #defines) also facilitates replacing the
version number coded into the Makefile with a Git description of the
working tree where appropriate.  This is done with GNU Make-specific
methods that don't interfere with "make: Nothing to be done for 'all'"
when nothing has changed.