Keep the makefile non-recursive
Write out prerequisites for test/test-vcf-api* (and add the executable
to .gitignore) so make can keep it up to date correctly.

Recursive makefiles are to be avoided because it's just not feasible to
get dependencies correct with them.  HTSlib now has correct dependency
information in its makefile, and it will be helpful to keep it that way.