Add hseek() and htell() test cases
Test the interaction between buffering and seeking (which requires
flushing or discarding buffers), and that htell()'s reported position
is kept in sync in spite of buffering.

It remains to add tests for interleaved reading and writing on files
opened for update, but at present no htslib code uses that anyway.