BGZF worker threads do no I/O themselves
Replace worker_t::fp with just the needed item (compress_level), to
ensure that worker threads continue to not do I/O on the BGZF file
themselves.  The upcoming hFILE is (currently) not thread-safe for
multiple threads working on the same hFILE.

Also make BGZF::mt a pointer to the correct incomplete struct rather
than a void*, so casts are no longer needed with it.