Enable warnings only if -Wno-unused-result works
While __attribute__((warn_unused_result)) was introduced in GCC 3.4,
the ability to ignore it with -Wno-unused-result was only introduced
in GCC 4.5 (and clang 3.0).  So activate these warnings for hclose() etc
only if we have the ability to temporarily ignore them conveniently.

(In particular, there was no way to shut it up with the system compiler
on Mac OS X Lion, llvm-gcc 4.2.1.)