Added missing return code checks.  Fixed 'x' mode of mfreopen.
Added missing checks for system call return codes.  Mainly malloc/realloc
but also i/o calls in mfflush.

Removed pointless check to see if the underlying file is seekable in
mfreopen when 'x' is present in the mode string.  The whole point of 'x'
was to turn off a seek in mfflush, so it doesn't matter if the file is
seekable or not.

(Code change added to io_lib by Rob Davies.)