Add sam_itr_*() dealing with BAM and CRAM iterators interchangeably
For CRAM, both indexing and restriction of iteration range are done
within the cram_fd, so we create a fake hts_idx_t and a dummy hts_itr_t
that bypasses the BGZF-focussed bulk of hts_itr_t and hts_itr_next().

This is a hack.  Ideally these functions would also work with tabixed
bgzipped SAM files via a tbx_t, but it's impractical to shoehorn that
into this hack due to the various separate extra metadata items needed.

The right way to do all this would be to make hts_idx_t and hts_itr_t
both opaque types with separate bgzf, cram, and tabix concrete versions
behind them, but that would be a larger change.