Add 'a' (append) mode letter; propagate mode letters
For bgzf_[dh]open() and hts_open(), treat 'a' the same as 'w';
the underlying hopen() will treat it differently and open the file
in append mode.  So that hopen() can do that (and so that it gets
to see other eventual mode letters that it will be interested in),
propagate _mode_ itself to underlying *_open() calls rather than
slicing it to just 'r' or 'w'.

It has been verified that there is (currently) no conflict amongst
the mode letters acted upon by different *_open() functions at their
different levels.