Add "data:blahblah" in-memory hFILE backend (read-only)
Currently read-only; implementing writing would be useful but a little
more complex.

This is akin to the data URI scheme (RFC 2397) but without the leading
MIME-type etc.  We may want to change this to require (or perhaps ignore
if present) a leading comma, which would make for the same syntax as the
URI scheme.