Fix sam_format1 failure on lines ending in integer/float tags and issue #63
This fixes a bug in sam_format1 where it gets a check to see if it has gone
off the end of the bam1_t data structure wrong when writing out optional
tags.  This affects any line ending in an integer or floating-point tag.

Also fixes __skip_tag so that it works properly with floating point tags
(issue #63).