Fixed REF_PATH to handle URLs better.
Without a REF_PATH it now defaults to using the EBI's MD5 service.
With a REF_PATH it now parses it in a friendly fashion. The URL= part
is now optional, instead auto-detecting http: and ftp: (no https yet).
Also even when using : as a path separator, when faced with a URL it
knows the syntax and treats the colons in http://host:80 as part of
the URL and not the separator. The old :: escape still works.