Added a range coder (order 0 and 1) plus support for LZMA if compiled in (no option for this at present).
Internally the choice of compression method is now a bit field
listing the various methods suitable.  To aid this, ZLIB_RLE is its
own method internally (although just ZLIB in the file format as it is
in the same ZLIB byte stream format), allowing better for auto-tuning
between fill ZLIB and RLE only variants.

This means we can specify as many encoding methods as we support and
get the code to auto-tune to the best methods available.