Bug fix of use of TAG_ID macro and k vs key in hash for auxiliary headers.
Updated the encoding of auxiliary data itself to agree with the
headers (and with the latest Scramble updates).

Bug fix to encoding the 'd' (double) aux data type.  We've never
tested
this as it's not even in the SAM spec, but I added it anyway as
samtools seems to include it.