Modified the thread pool to use as few threads as possible. The intention is that when given, say, 16 threads but being I/O bound such that 12 cores is enough to keep up with the I/O then we have 12 threads kept busy and 4 totally idle.
This is implemented by dispatching jobs on lower thread IDs in
preference and maintaining a queue of pending jobs in proportion to
the number of threads currently running, rather than instantly
dispatching.

Removed uninitialised memory behaviour (worse, freeing an
uninitialised pointer). This triggered crashes on SPARC/Solaris when
testing in multi-threading, but bizarrely never fired on intel/linux
not even under valgrind.