Reverting 9e844e0be26875d29b6e2c853c2b4d02c4f1faa4 and rewriting in a better manner.
Passing +1 to the extra_length in bam_construct_seq causes the length
of data in the bam structure to be +1, and then causes bam decoding to
have issues in some cases.

The correct solution made here is simply to remove the nul terminating
of the auxiliary block.  It's a hang-over from the io_lib days and
isn't necessary.