Fixed an incorrect EOF-style case when seeking multiple times.
When asking to fetch data for a specific chr:range, if the current
container entirely fits within the range but the next container is
outside the range then it sets the fd->ooc (out of containers) flag to
prevent subsequent containers from being decoded, to mark the end of
the sub-query.

However this then causes a subsequent cram_seek to also fail as the
ooc flag is still set. We now clear this on every new seek.