Two changes to cram_encode_container() related to efficient encoding of name-sorted data.
Firstly, if we have multiple reference sequences packed into one
container, only pre-populate the relevant fd->refs[] array entries if
we're doing reference based encoding.

Secondly, a bug in this same code was that we pre-cached without
incrementing the reference count (but did decrement it at the end of
the function).  The practical upshot of this is that the reference
count added due to fd->shared_ref (enabled when doing multi-ref
slices) was decremented and consequentially caused reference sequences
to be reloaded every container instead of cached between them.