Add configure.ac script
Use autoconf to generate a configure script.  Checks that zlib development
files are available, and emits a hopefully informative error if not.

The inclusion of config.mk is carefully placed so that it can add to
prerequisites as appropriate when later --enable-xxx/--with-yyy options
need to add extra libhts objects and rules.

Rewrite INSTALL to describe configure usage.