Bug fix to refs_from_header().
The patch to hts_set_fai_filename() caused a new failure mode in this
function. If given a .fai file with sequence records that are not in
the same order as the @SQ lines in a SAM header, the code to parse the
@SQ header and potentially add new records that aren't in the original
.fai file was failing, returning -1.

We now do a better job of merging the references together.