Detect file format versions in hts_detect_format()
Adds version fields to htsFormat and removes bcfv1 from htsExactFormat
(even though it's really a separate format, this way is less confusing).
As htsExactFormat and htsFormat have not yet appeared in an HTSlib release,
changing these types' layout now is not a compatibility problem.  But soon
it will be, so also adds compression_level and [format-]specific fields, to
be used by an upcoming hopen(filename, "w", &format) open-for-write call.

hts_format_description() now builds a more accurate description string
based on all the format fields.  The string it returns must now be freed
by the caller -- cf htsfile.c, to which we also add --version and fix -?.

(Addresses #149, include compression info in htsfile description output.)