Include unistd.h & exec bash instead of /bin/bash
Explicitly include unistd.h in cram_io.c, as it declares several
(fsync(), getcwd(), and access()) and the symbolic constant R_OK, and
may not be included by any of the other included headers in cram_io.c.

In test.pl, use exec('bash', ...) instead of exec('/bin/bash', ...), as
bash isn't guaranteed to be in /bin/ on all platforms (e.g., FreeBSD),
or the user may want to specify a different version of bash via their
PATH environment variable.