Fix for unmapped reads with "*" sequence.  These were defaulting to a sequence length of -1 instead of 0 (it's unknown really), causing an attempt to extract -1 bytes.
Also fixed NM and MD handling for sequence "*".  In this case we
cannot derive it as we have nothing to compare against, so we store
any values verbatim.  The same bug has been fixed for CRAM_OPT_NO_REF
mode.

The change is quite complex as it now also checks which of MD and/or
NM tags are already present (now that we case sometimes omit them
verbatim) in order to avoid creating duplicate tags.