Convert relative hseek(SEEK_CUR) offsets to absolute SEEK_SET positions
It turns out to be preferable to convert to SEEK_SET rather than to adjust
the SEEK_CUR offset, as some backends (e.g. the upcoming libcurl backend)
may not track their current physical position and would prefer not to
implement SEEK_CUR.  With this change, we send only SEEK_SET or SEEK_END
to the backend.

Compare glibc's fseek(), which uses lseek(SEEK_SET) for everything except
in unusual circumstances.

(Ideally we'd check for imminent overflow without actually overflowing, but
that's non-trivial in the absence of an OFF_T_MAX or a known file size.)

Refixes #152.