Added C/Java cross-validation script for CRAM.
This isn't yet part of our automated test harness, but it is useful to
have a copy in git for periodic manual checking.  Also improved the
compare_sam.pl with an option for comparing B-type auxiliary tags as
the Java code changes sign and also changes H to B.  I am unsure if
this is valid, but it helps if we can hide the known differences in
order to spot unknown ones.