Bug fix to the CRC32 checking for Java/C integration.
The block and container header CRCs check sum variable sized data.
The old method was to decode the structure and then reencode to a
block of memory that we can compute the CRC on.  Unfortunately there
is more than one valid encoding for the same numerical integer value,
and even more unfortunately Java and C implementations differ.  This
gives rise to false CRC failures.

The new code computes the CRC as it goes using the actual bytes being
decoded.  This isn't as quick, but the measured difference in speed is
under 1%.