Add and use safe_itf8_get() to avoid more buffer over-runs.
Add new static inline safe_itf8_get, to replace the old itf8_get macro.
It takes a pointer to the end of the buffer so it can test if it has run
out of data.  If it fails to read a complete integer it returns 0,
otherwise it returns the number of inout bytes consumed, as for itf8_get.

Change lots of uses of itf8_get() to safe_itf8_get().  Also add some
other checks for running out of data.

Make cram_decode_slice_header uncompress the slice header block if it
isn't of type RAW.  This goes beyond what the CRAM specification says,
but there's no real reason why it can't be compressed.

All this fixes several bugs found during "american fuzzy lop" fuzz testing.