Fixed error handling in CRAM.
A bit of a howler: closing a file when we're not at EOF is not an
error!  We can legitimately open a file, read 10 of 100 records, and
then close.

Improved sam_read1 return values so that it can distinguish between
eof (-1) and error (-2 or lower).  This is done by using cram_eof and
checking that a -1 return value is due to eof rather than something
else.

These two combined essentially move the onus of error detection away
from hts_close to sam_read1, as per bam I/O.