Removed quality values from the CRAM_SEQ macro and added a CRAM_QUAL one.
This is an optimisation so that, for example, bam2depth, doesn't
decode quality values unnecessarily.  Note that due to the existance
of some combined feature codes such as 'A' (base and score) and 'B'
(read base) that access both BA and QS data series, it may be that
decoding a sequence explicitly requires fetching from the quality
stream too.

In this situation, there is explicit ds&CRAM_BA and ds&CRAM_QS so only
the half of the data series needed will be decoded.