[faidx] Warn if duplicate sequence names are encountered
Add only the first sequence of each name to the faidx_t and .fai file.
An alternative would be to uniquify subsequent duplicates by changing
their names to "<NAME>_<N>" or something, but there's a risk of introducing
further clashes -- better to leave the user to sort it out.

Fixes samtools/samtools#380.  An upcoming samtools documentation change
will note that subsequences retrieved are from the *first* instance of
duplicate-named reference sequences.