More fuzz-detected BAM file input checks
(Squashed from PR #217, and partly tweaked by the committer.)

* Add missing test for running out of data to sam_format1.

Makes the code to handle B-type tags check that it has at least 4 bytes
available before reading the number of items.

* Ensure c->l_qname >= 1 in bam_read1.

qname must have at least a NUL in it, so the minimum possible length is 1.
Prevents a possible wrap-around in sam_format1 which uses the value
c->l_qname-1.

* Fix test for enough bytes left when converting B-type tags in sam_format1

The check wasn't strict enough as it did not take the size of the array
elements into account.

The test for enough bytes in sam_format1 has been arranged so that it
doesn't try to work out n * sub_type_size.  This avoids the possibility
of overflowing an int32_t.