Ensure names are present for @SQ/@RG/@PG lines in sam_hdr_update_hashes
sam_hdr_update_hashes goes through the tags on a header line that it's
interested in, filling in values in a SAM_hdr_tag struct as it goes.
One of these is name, which is set to a copy of the SN tag for @SQ lines
and ID for @RG and @PG lines.  These must be present according to the SAM
specification, so ensure that name has been set after going through all
the tags for the line.  This means anything using the name field (notably
refs_from_header) can assume that it isn't NULL.

Fixes a crash found by the Americal Fuzzy Lop fuzz tester.

Also fix some potential memory leaks that could happen if realloc
fails in this function.