Fix test for enough bytes left when converting B-type tags in sam_format1
The check wasn't strict enough as it did not take the size of the array
elements into account.

Adds aux_sub_type2size() which is a copy of aux_type2size() specialized for
the case of B-type tags.  For completeness, update swap_data() to use
this new function as well.

The test for enough bytes in sam_format1 has been arranged so that it
doesn't try to work out n * sub_type_size.  This avoids the possibility
of overflowing an int32_t.