Fix handling of zero-bit beta codec case.
Using the beta codec with nbits=0 is valid, but confuses
cram_not_enough_bits() as it's possible to have no input data in
this case so cram_not_enough_bits thinks it has run out.

Move the calls to cram_not_enough_bits inside checks that c->beta.nbits
is non-zero to avoid the problem.