Specify signedness for bit-fields explicitly
Plain int bit-fields technically have indeterminate signedness, and in
particular 1-bit signed bit-fields are unexpectedly entirely sign bit,
producing warnings on "s.foo = 1; /* should be -1 */" assignments with
gcc -Wsign-conversion and copious warnings on the struct definitions in
the headers with NVIDIA's nvcc.  Fixes #225, hat tip @nsubtil.

The changed hFILE fields are in (semi-)opaque structs so changing them
brings no compatibility concerns.  Some BGZF fields are used by client
code, so we are careful not to change the struct's layout.  The 2-bit
bit-fields are in fact used as booleans, so are presumably 2 bits as
an inferior sign bit workaround.  We've left them with the same sizes
so as not to change their layout.

(There remains one plain int bit-field in htslib/kseq.h.  We may fix
that later in conjunction with klib.)