Add generic plugin infrastructure
PLUGIN_PATH is set in the makefile (rather than config.h) as it likely
contains installation directory variables, which need to be expanded
and ought to be alterable at "make" time.

Moved "ifeq $(PLATFORM)" etc to below "include config.mk" so that it
can act on the config.mk-modified $(PLUGIN_OBJS).

This way of adding -rdynamic is something of a hack, but it seems
desirable to do this alongside -ldl.  (TODO) The eventual plugins
themselves don't want to be linked with -rdynamic -ldl.

Document new configure options (including --enable-libcurl already added).