Remove abort() and improve multi-threaded worker error handling
Get rid of the calls to abort() and return error codes instead.

Make worker_aux() give up in a cleaner fashion if it spots an error.  In
particular, after setting its error flag and tidying up, it ensures that
w->mt->proc_cnt is incremented so that mt_flush_queue() won't get stuck
in an endless loop waiting for it to finish.  It then returns 0 so it
will be called again, where it can wait for done to be signalled in
the normal way.

Ensure that mt_flush_queue() only writes data if no errors were reported
from the threads (otherwise the output would have missing blocks).