Make plugin handling API slightly less generic
For HTSlib, we will have the convention that each category of plugin has
a well-known symbol that the plugin must provide, and load_plugin() both
opens the plugin and finds that symbol.  This simplifies plugin-loading
code for the usual case of having one extern symbol, and plugin_sym() is
available for any subsequent extern symbols.

We consider failure in plugin loading to be a case of missing optional
functionality, so also put the dynamic loader error reporting here at a
verbose log level (hts_verbose >= 4).