Work around an iRODS bug(?) regarding checksum and file updates.
If we open, read/write and close a file then the irods close call
takes an exceptionally long time (possibly 10min+), apparently due to
repeated attempts to create and verify a replica.

By forcibly recomputing the checksum prior to closing, the replication
works and verifies so it exits much quicker.