Recognise several tab-separated text file formats
Enlarge the buffers in decompress_peek() and hts_detect_format() so
that an entire line or two can be peeked at. Parse a tab-separated line
and match the column types (but allowing for the possibility that we've
only peeked an incomplete prefix of the line) to try to differentiate
SAM, CRAI, and BED, with others potentially to follow.

Note that avoiding false negatives is vital: if hts_detect_format()
fails to recognise a SAM file as such, then hts_open() will refuse to
open it.

Fixes #200.

Relax htslib/hts.h's hts_open() documentation to (re-)bless its use for
opening arbitrary text files to be read line-by-line via hts_getline().
Such files will now be detected as text_format or bed (rather than the
previous "everything textual was sam"); update tabix.c accordingly.