Constify hts_open_format() parameter and don't parse filename
Don't override mode based on the filename extension, as this level of
intelligence is not desirable by default in a low-level open() function.

Constify fmt parameter, as it is naturally input-only (if the detailed
format as opened is desired, it should be queried via hts_get_format()).

(Note that hts_open_format() documentation needs updating.)