Change rANS O1 codecs to allocate large arrays on the heap.
rans_compress_O1 needed about 1.25 Mb of stack for a couple of 256x256
arrays.  This caused problems on MacOS X when running in threaded mode
due to running out of per-thread stack.  The default stack size for
threads on MacOS X is 512 K.  For details, see:
https://developer.apple.com/library/mac/qa/qa1419/_index.html

This changes rans_compress_O1 and rans_uncompress_O1 to use malloc
and calloc to allocate the arrays instead of using the stack.