Improved the efficiency of reference loading again.
Commit 47da392939b43c9761454278de5f457119c94297 included a change to
use bgzf_open and read a local file rather than relying on
open_path_mfile if it was found via REF_PATH as a local on-disk file.

It turns out this is only beneficial if we don't have HAVE_MMAP
defined, which by default isn't on in the Makefile (but likely will be
if using ./configure).  With mmap the mfopen becomes a trivial
operation.