Reduced memory for CRAM decoding.
This is particularly important during merging or multi-file mpileup.
The changes involve reducing the block size in pooled_alloc (saves
SAM header memory), freeing loaded disk-blocks immediately after
decoding rather than when freeing the slice, and reducing the memory
usage of the decoded arrays to the exact size needed rather than an
auto-growing array (which at worst case is 50% larger).

It also attempts to automatically predict the size of the sequence,
quality and read-name memory blocks instead of auto-growing.  This is
both a performance improvement and potentially a reduction in heap
fragmentation.