Fix handling of BCF string dictionaries
Gaps in IDX blocks were causing incorrect tags on output.

The allocated size of the dictionary block in use is added to the
bcf_hdr_t typedef.  It has been added to the end of the struct to
preserve the structure layout, and a note added recommending allocation
via bcf_hdr_init().

Fixes samtools/bcftools#317.