Fixes bug that appeared with commit ca1bcefe.
The commit fails to check whether we have requested the AP data series
to be decoded, so causes a failure in samtools flagstat.  However this
failure then triggers another bug when running in multithreaded mode;
it attempts to flush any remaining blocks during close, even when
decoding produced an error.  This crashes due to an incomplete
container.

Fixes #265