Define bcf_empty() to match its declaration in vcf.h
Previously we declared bcf_empty() but defined bcf_empty1(), so the
function in the public header was unusable (and there was no public
declaration for the function that did exist).

Correct this by adding a bcf_empty1() alias macro for consistency with
similar functions, and defining bcf_empty() to match the API function
declaration.  Deobfuscate bcf_init() and bcf_destroy() definitions,
which previously appeared to be defining bcf_init1()/bcf_destroy1()
but were not as the alias macros act on these definitions too.

This constitutes adding bcf_empty() to the API.  There is no ABI
breakage due to removing the bcf_empty1() symbol from the library,
as bcf_empty1() was not declared in the public header files so is
not part of the API.