Fix plugin installation rules
With plugins disabled, 'test -n ""' is false but we need to ensure that
the exit status of the entire recipe line is success.  D'oh.  Fixes #296.

It might be nicer to have new installdirs-plugins/install-plugins rules
that are activated by config.mk, but that too would need to handle the
case of --enable-plugins with no plugins actually being built -- which
would make it not so nice.