Add read error check to fai_load
fai_load was invoking fai_read which did not check for ferror() on
read.  (Fixes #315.)  Add some system error reporting to fai_load.

This removes fai_read() as a public symbol in libhts.so, but this
function does not appear in a public header file so is not part of
the HTSlib API.  So there are no ABI implications of this removal.