Fixes possible memory leaks, NULL dereferences and integer overflow.
Fixes memory leaks in bgzf.c detected by cppcheck, mainly when bailing
out of functions on error.  At the same time put in some missing checks
for failed memory allocation and fix a possible integer overflow in
bgzf_index_dump where it uses a value read from the index file in a
malloc.

The tests for fread reading the correct number of bytes in bgzf_index_dump
are moved slightly so it will stop faster if it hits EOF.