Add checks for malloc, ferror to fai_read, fai_load and fai_insert_index
Fix a memory leak found by cppcheck-1.72 in fai_load.  At the same time,
fix various other places where allocations are not checked or memory
could be leaked, notably in fai_insert_index which is also changed to
return non-zero if it fails.

Add a check on ferror in fai_read to catch file reading errors.