Detect non-numeric characters in numeric format fields
Stop reading on characters-other-than-comma rather than when we see
the expected colon or end-of-field.  Fixes #321, fixes #322, fixes #323.

At present prints an error message and aborts parsing.
TODO We may need to add a new BCF_ERR_* value to |= into v->errcode,
and we need to consider whether plain "return -1" is appropriate.
(Adding a setting to reduce this to a warning would facilitate adding
the reported example VCF records as test cases in test/formatcols.vcf.)