Adds HTS_OPT_BLOCK_SIZE support for SAM/BAM/CRAM.
Allow the size of the internal hFILE buffer to be changed.  This may
be useful for fine tuning I/O speed on filesystems that don't report
an optimal block size.

It's possible to shrink the buffer, but only if the buffer does not
contain data that would be lost after the resize.  If it does, a
warning will be printed and the buffer will be left at the existing
size.