Rename kpa_glocal() to probaln_glocal() and add to htslib/hts.h
There are only four declarations to be added to the API, so declare them
in htslib/hts.h rather than their own header file.  The kpa_ prefix was
obscure and misleading (as this has always been local to samtools and
never part of klib), so change the prefix to probaln_ now that this is
being made public.  Rename kprobaln.c to probaln.c similarly.

(TODO) Add documentation in hts.h, based on the explanation in probaln.c.