Added a safe_ltf8_decode function to go along with the itf8 variant.
Also fixed a potential speed issue when doing multi-threaded
encoding.  This was fixed in io_lib svn r3956, but for some reason not
duplicated here (unless there was a reason it didn't count, which
escapes me if so).

Plus a couple harmless/tiny code tweaks to improve diffing against
io_lib.