Added a BASES_PER_SLICE hts option.
This is used in CRAM to set an upper limit on the number of bases per
slice in addition to the sequences per slice.  The default value is 5
million, equiv to 10,000 reads (the default seqs per slice) each 500bp
long.  Note adjusting seqs_per_slice does not automatically scale
bases_per_slice.

The reason for this is to prevent excessive memory usage when encoding
or decoding long read technology data (eg PacBio or ONT).