VCF GT parsing fix
Also allow for omitted trailing GT FORMAT fields.  Compare 5c65920de10f3068802b6d4df1be86cfcd2ce5cd.
This is not valid VCF (in particular, because GT when present must
be the first sub-field), but we shouldn't write out-of-bounds if we
encounter it.  Hat tip @cryptoad.

Add test case, in which the non-last GT fields are misprinted as "2,4"
because the gt_i code in vcf_format() cannot handle multiple GT sub-fields
(which is invalid anyway).  Integer arrays and GT are represented the same
way in memory, so fortunately this does not have OOB reading implications.