Sped up rANS decoders by 6% (O0) to 10% (O1).
This is largely achieved by reducing the size of the ari_decoder
struct.  It had a lot of temporary data there which was needed to
create the lookup tables but not needed during decoding.

The R[] array is also no longer calloc/memset as it's entirely defined
by the decoder and can therefore start off uninitialised.

Also slightly reordered the decoding code as this seems to give a
small improvement to speed.

The aggregate change to samtools as a whole when decoding CRAM is
around 5%, mostly achieved through a reduction in frontend cycles idle
from 36 to 35%.   This is still too high for my liking though.